seperated gas_curve class from file to increase readability. Getters and setters added
